+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - `debug` config flag
 
 ### Changed
+- Primary authentication is less hacky.
 
 ### Fixed