+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## Unreleased
+## [0.3.0](https://github.com/flowerysong/pam_okta_auth/releases/tag/v0.3.0) - 2025-07-07
 
 ### Added
 - `debug` config flag
@@ -11,8 +11,6 @@ All notable changes to this project will be documented in this file.
 ### Changed
 - Primary authentication is less hacky.
 - HTTP requests send a custom User-Agent header instead of the library's.
-
-### Fixed
 
 ## [0.2.0](https://github.com/flowerysong/pam_okta_auth/releases/tag/v0.2.0) - 2025-06-27